Implement the prototype of the master note system
The progress is not yet fully implemented but it is getting there. Also the subject has been moved to its own file.
